Add files via upload
converShoes
Converse is an online shoe store website designed to provide a seamless and interactive user experience. Built with HTML, CSS, and JavaScript, it features responsive design and modern UI/UX principles. The site utilizes Locomotive Scroll for smooth scrolling effects and GSAP for dynamic animations, enhancing the visual appeal and user engagement. The project showcases a collection of shoes with detailed descriptions and interactive elements, making it both functional and visually appealing. Perfect for those looking to explore the intersection of design and interactivity in web development.